Introducing dictionary

+++ Board.cpp
@@ -24,8 +24,8 @@ Field *Board::get(int x, int y) {
 }
 
 void Board::printBoard(std::wostream& os) const {
-	for (int i = 0; i < fields.size(); ++i) {
-		for (int j = 0; j < fields[0].size(); ++j) {
+	for (size_t i = 0; i < fields.size(); ++i) {
+		for (size_t j = 0; j < fields[0].size(); ++j) {
 			os << *fields[i][j];
 		}
 		os << std::endl;
